Rename files to remove redundant naming
